pyproject.toml
[tool.poetry]
name = "sdss-lvmscp"
version = "0.1.2"
description = "Simple Template package for creating SDSS Python projects"
authors = ["Changgon Kim <changgonkim@khu.ac.kr>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/lvmscp"
repository = "https://github.com/sdss/lvmscp"
documentation = "https://sdss-lvmscp.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3.8",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "lvmscp", from = "python" }
]
include = ["python/lvmscp/etc/*", "cextern/**/*", "LICENSE.md"]

[tool.poetry.build]
script = "build.py"
generate-setup-file = false

[tool.poetry.scripts]
lvmscp = "lvmscp.__main__:lvmscp"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
sdss-tree = ">=2.15.2"
sdss-access = ">=0.2.3"
sdsstools = ">=0.4.0"
Sphinx = {version="^3.0.0", optional=true}
sphinx_bootstrap_theme = { version = ">=0.4.12", optional=true}
astropy = "^4.2.1"
click = "^8.0.0"
click-default-group = "^1.2.2"
sdss-clu = "^1.0.0"
daemonocle = "^1.2.3"
black = "^21.7b0"
sphinx-click = "^3.0.1"
sdss-cluplus = "^0.0.14"

[tool.poetry.dev-dependencies]
ipython = ">=7.11.0"
matplotlib = ">=3.1.1"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
isort = ">=4.3.21"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
rstcheck = "^3.3.1"

[tool.poetry.extras]
docs = ["Sphinx", "sphinx_bootstrap_theme"]

[tool.isort]
line_length = 79
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "lvmscp"
known_sdss = ["sdsstools"]
balanced_wrapping = true
include_trailing_comma = false
lines_after_imports = 2
use_parentheses = true

[tool.pytest.ini_options]
addopts = "--cov lvmscp --cov-report xml --cov-report html --cov-report term"

[tool.coverage.run]
branch = true
include = ["python/lvmscp/*"]
omit = [
    "python/lvmscp/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
]

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=49.6.0"]
build-backend = "poetry.core.masonry.api"
